+++ CHANGELOG.md
@@ -10,6 +10,15 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 <!-- TODO: change `@added unknown` tag due to the actual version -->
 <!--       during new scriptlets or redirects releasing -->
 
+## [Unreleased]
+
+### Added
+
+- ability in `prevent-addEventListener` scriptlet to match specific element [#480]
+
+[Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.1.6...HEAD
+[#480]: https://github.com/AdguardTeam/Scriptlets/issues/480
+
 ## [v2.1.6] - 2025-03-06
 
 ### Fixed